Use "www-data" as web_user for all dev instances.
This is coordination with a change on the Packer side to stop skipping the
"users" role (via --skip-tags). With that role now running on Vagrant hosts,
we also need to supply a sudoer password (which is also the user account
password) so that we don't overwrite the default password of "vagrant".